+++ core/src/Contour.cpp
@@ -224,12 +224,12 @@ void Contour::findClosestPoints(TGraph* g1, TGraph* g2, int& i1, int& i2) {
   //   x1, y1, i1, x2, y2, i2);
 }
 
-// float Contour::getXBoundary(float p1x, float p1y, float p2x, float p2y, float ymax)
+// double Contour::getXBoundary(double p1x, double p1y, double p2x, double p2y, double ymax)
 //{
 // return p1x + (ymax-p1y)/(p2y-p1y)*(p2x-p1x);
 //}
 
-// float Contour::getYBoundary(float p1x, float p1y, float p2x, float p2y, float xmax)
+// double Contour::getYBoundary(double p1x, double p1y, double p2x, double p2y, double xmax)
 //{
 // return p1y + (xmax-p1x)/(p2x-p1x)*(p2y-p1y);
 //}
@@ -242,13 +242,13 @@ void Contour::findClosestPoints(TGraph* g1, TGraph* g2, int& i1, int& i2) {
 /// \param hCL - a histogram defining the boundaries
 ///
 void Contour::magneticBoundaries(vector<TGraph*>& contours, const TH2F* hCL) {
-  float magneticRange = 0.75;
-  float xmin = hCL->GetXaxis()->GetXmin();
-  float xmax = hCL->GetXaxis()->GetXmax();
-  float ymin = hCL->GetYaxis()->GetXmin();
-  float ymax = hCL->GetYaxis()->GetXmax();
-  float xbinwidth = hCL->GetXaxis()->GetBinWidth(1);
-  float ybinwidth = hCL->GetYaxis()->GetBinWidth(1);
+  double magneticRange = 0.75;
+  double xmin = hCL->GetXaxis()->GetXmin();
+  double xmax = hCL->GetXaxis()->GetXmax();
+  double ymin = hCL->GetYaxis()->GetXmin();
+  double ymax = hCL->GetYaxis()->GetXmax();
+  double xbinwidth = hCL->GetXaxis()->GetBinWidth(1);
+  double ybinwidth = hCL->GetYaxis()->GetBinWidth(1);
   Double_t pointx, pointy;
   for (auto contour : contours) {
     auto g = dynamic_cast<TGraph*>(contour);
@@ -275,7 +275,7 @@ void Contour::magneticBoundaries(const TH2F* hCL) {
 ///
 /// \param percent - 0% means intransparent
 ///
-void Contour::setTransparency(float percent) {
+void Contour::setTransparency(double percent) {
   if (!(0. <= percent && percent <= 1.)) {
     cout << "Contour::setTransparency() : ERROR : percent not in [0,1]. Skipping." << endl;
     return;
